# docs/yaml/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Back End Demonstration API
  description: A simple blog API.
  contact: 
    name: Tse Ho Fai
    email: 227021367@stu.vtc.edu.hk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:10888/'
    description: Development server
tags:
  - name: Users
    description: API endpoints for user management.
  - name: Cats
    description: Access and perform CRUD operations on cat entries.
paths:
  /cats:
    get:
      tags:
        - Cats
      description:
        Get all cat
      responses:
        '200':
          description: Return exisiting cat list from database
        '404':
          description: No data from database

  /cats/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Id of the cat
    get:
      tags:
        - Cats
      description:
        Get cat by id.
      responses:
        '200':
          description: View of an cat record.
        '404':
          description: Cat was not found in the DB.
    post:
      security:
        - basicAuth: []    
      tags:
        - Cats
      description:
        Create a new cat
      requestBody:
        description: A new cat data
        required: true
        content:
          application/json:
            schema:
              $ref: ./cat.json
        responses:
          '201':
            description: Cats created
    put:
      security:
        - basicAuth: []
      tags:
        - Cats
      description: Update cat by ID.
      content:
        application/json:
          schema:
            $ref: ./cat.json
      responses:
        '200':
          description: View of an cat record.
        '404':
          description: Cat was not found in the DB.
    delete:
      tags:
        - Cats
      description:
        Delete cat by ID.
      responses:
        '200':
          description: View of an cat record.
        '404':
          description: Cat was not found in the DB.

  /users:
    post:  
      tags:
        - Users
      description:
        User Register
      requestBody:
        description: A new cat data
        required: true
        content:
          application/json:
            schema:
              $ref: ./user.json
      responses:
        '201':
          description: Register successfully
        '500':
          description: Insert Data failed   
          
  /users/login:
    post:
      tags:
        - Users
      description:
        User Login
      requestBody:
        content:
          application/json:
            example:
              username: Bob
              password: 123
      responses:
        '200':
          description: Login successfully
        '500':
          description: Login Fail  